Now I want to do a couple of other cleanup things. First: whenever we need to read information off of the request that the client is sending us, we're going to use this same json_decode for the content:
 31 lines  src/KnpU/CodeBattle/Controller/Api/TokenController.php
... lines 1 - 16
  public function newAction(Request {dollar_sign}request)
... lines 18 - 20
    {dollar_sign}data = json_decode({dollar_sign}request->getContent(), true);
... lines 22 - 28
  }
... lines 30 - 31
In fact, we use this in ProgrammerController inside of the handleRequest method.:

 168 lines  src/KnpU/CodeBattle/Controller/Api/ProgrammerController.php
... lines 1 - 123
  private function handleRequest(Request {dollar_sign}request, Programmer {dollar_sign}programmer){
    {dollar_sign}data = json_decode({dollar_sign}request->getContent(), true);
... lines 127 - 154
  }
... lines 156 - 168
So we have the same logic duplicated there. Let's centralize this by putting it into our BaseController. Open this up and create a new protected function at the bottom called decodeRequestBodyIntoParameters. I know, really short name. And we'll take in a Request object as an argument. And at first this is going to be really simple. We can go to the TokenController, grab this json_decode line, go back to BaseController and return it.

So now in the TokenController, {dollar_sign}data = {dollar_sign}this->decodeRequestBodyIntoParameters({dollar_sign}request) and we've got it.

So just to be sure, let's go back and run our entire feature:

php vendor/bin/behat features/api/token.feature