Consistently Erroring on Invalid JSON
Everything still passes so we're good. So, why did we do this? Back in ProgrammerController, when we decoded the body in handleRequest, we also checked to see if maybe the json that was sent to us had a bad format. If the JSON is bad, then json_decode is going to return null which is what we're checking for here.

So let's move that into our new BaseController method, because that is a really nice check. And then it's creating an ApiProblem and throwing an ApiProblemException so we can have that really nice consistent format. We just need to add the use statements for both of these.

Perfect. Let's rerun those again to make sure things are happy ... and they are!