Don't Blow up on an Empty Request Body!
One other little detail here was that if the request body was blank this was going to blow up with an invalid request body format because json_decode is going to return null. Now technically sending a blank request is not invalid json so I don't want to blow up in that way. This doesn't affect anything now but it's planning for the future. So if ! {dollar_sign}request->getContent(), then just set  to an array. Else, we'll do all of our logic down here that actually decodes the json.

And just to make sure we didn't screw anything up, we'll rerun the tests.

php vendor/bin/behat features/api/token.feature.