A ParameterBag Makes Life Nicer One last little thing that is going to make our code even easier to deal with. Back in TokenController, because the decodeRequestBodyIntoParameters returns an array, we need to code a bit more defensively here. What if they don't actually send a notes key, we don't want some sort of PHP error.
The horror!

And that's not that big of a deal but it's kind of annoying and error prone. So instead, in our new function I want to return a different type of object called a ParameterBag.

This comes from a component inside of Symfony that Silex uses. It's an object but it acts just like an array with some extra nice methods. Let me show you what I mean, back in TokenController instead of using it like an array we can now say {dollar_sign}data->get() and if that key doesn't exist it's not going to throw some bad index warning. We can also use the second argument as the default value