Use decodeRequestBodyIntoParameters in all the Places
We have this really nice new function inside of our BaseController and I also want to take advantage of it inside of our ProgrammerController. We're going down to handleRequest and now we can just say {dollar_sign}data = {dollar_sign}this->decodeRequestBodyIntoParameters() and pass the {dollar_sign}request object.

Next, this big if block is no longer needed. And now because data is an object instead of an array, we need to update our two or three usages of it down here, which is going to make things simpler. So instead of using the isset function, we can say if(! {dollar_sign}data->has({dollar_sign}property)) because that's one of the methods on that ParameterBag object. And down here instead of having to code defensively using isset, we can just say {dollar_sign}data->get({dollar_sign}property). In fact let's just do this all in one line.

Lovely! Now that was a fairly fundamental change so there is a good chance that we broke something. So let's go back and run our entire programmer feature:

php vendor/bin/behat features/api/programmer.feature
Beautiful! We didn't actually break anything which is so good to know.