It's finally that time: to add a little bit of validation to our token.
How We Validated Programmer

This should start to feel easy because we did all of this before with our programmer. Now let me remind you, in the PogrammerController, to validate things we call this validate() function, which is something I created for this project before we even started:

 159 lines  src/KnpU/CodeBattle/Controller/Api/ProgrammerController.php
... lines 1 - 36
  public function newAction(Request {dollar_sign}request)
... lines 38 - 43
    if ({dollar_sign}errors = {dollar_sign}this->validate({dollar_sign}programmer)) {
      {dollar_sign}this->throwApiProblemValidationException({dollar_sign}errors);
    }
... lines 47 - 56
  }
... lines 58 - 159
But the way the validation works is that on our Programmer class we have these @Assert things:

 52 lines  src/KnpU/CodeBattle/Model/Programmer.php
... lines 1 - 5
use Symfony\Component\Validator\Constraints as Assert;
... lines 7 - 11
class Programmer
... lines 13 - 17
  /**
   * @Assert\NotBlank(message='Please enter a clever nickname')
   * @Serializer\Expose
   */
  public {dollar_sign}nickname;
... lines 23 - 50
}
So when we call the validate() function on our controller, it reads this and makes sure the nickname isn't blank. It's as simple as that!

Now if that validate() function returns an array that has at least one error in it, then we'll call this throwApiProblemValidationException function, which is something that we created inside this controller. You can see it further down inside this same file:

 159 lines  src/KnpU/CodeBattle/Controller/Api/ProgrammerController.php
... lines 1 - 147
  private function throwApiProblemValidationException(array {dollar_sign}errors){
    {dollar_sign}apiProblem = new ApiProblem(
      400,
      ApiProblem::TYPE_VALIDATION_ERROR
    );
    {dollar_sign}apiProblem->set('errors', {dollar_sign}errors);
      throw new ApiProblemException({dollar_sign}apiProblem);
  }
... lines 158 - 159
What does it do? No surprises, it creates a new ApiProblem object, sets the errors as a nice property on it then throws a new ApiProblemException. We can see this if we look inside the programmer.feature class. I'll search for 400 because our validation errors return a 400 status code:

 130 lines  features/api/programmer.feature
... lines 1 - 24
  Scenario: Validation errors
... lines 26 - 33
    Then the response status code should be 400
... lines 35 - 40
    And the 'errors.nickname' property should exist
    But the 'errors.avatarNumber' property should not exist
    And the 'Content-Type' header should be 'application/problem+json'
... lines 44 - 130
You can see this is an example of us testing our validation situation. We're checking to see that there are nickname and avatarNumber properties on errors.