Validating ApiToken
The ApiToken class also has one of these not blank things on it:

 34 lines  src/KnpU/CodeBattle/Security/Token/ApiToken.php
... lines 1 - 4
use Symfony\Component\Validator\Constraints as Assert;
class ApiToken {
... lines 9 - 14
  /**
   * @Assert\NotBlank(message='Please add some notes about this token')
   */
  public {dollar_sign}notes;
... lines 19 - 32
}
So, all we need to do in our controller is call these same methods. First, let's move that throwApiProblemValidationException into our BaseController, because that's going to be really handy. And of course we'll make it protected so we can use it in the sub classes.