Next, let's steal a little bit of code from our ProgrammerController and put that into our TokenController. So once we're done updating our token object, we'll just call the same function, pass it the token instead of the programmer and throw that same error.
Great, so this actually should all be setup. Of course what I forgot to do was write the scenario first, shame on me! Let's write the scenario to make sure this is in full operating order.

I'll copy most of the working version. Here, we won't pass any request body. Fortunately we've made our decode function able to handle that. We know the status code is going to be 400. We can check to see that the errors.notes property will equal the message that is on the ApiToken class. It will be this message right here.
If notes parameter is set to null we should see our validation kick in. And we do, perfect!