Let's login to the site. Don't forget, our tests like to mess with our database, so I'm going to delete the SQLite database file and it'll regenerate with some nice test data:
rm data/code_battles.sqlite
We'll login as ryan@knplabs.com password foo.

We already know that I'm able to create a programmer. And we even have some really nice API endpoints for this. The other part of the site is all about battles. If I click 'Start Battle', this is a list of projects that are in the database right now. If I dare to select one of those project, it starts an EPIC CODE Battle OF HISTORY between the programmer and the project and picks a winner.

A battle is another type of resource, but it can't be created yet in the API. Let's fix that!

New Battle Feature

Like my other resources, I already have a class that models this. You can see there's a programmer, a project, the outcome didProgrammerWin and it even stores the date it was fought and some extra notes:

26 lines  src/KnpU/CodeBattle/Model/Battle.php
<?php
namespace KnpU\CodeBattle\Model;
class Battle {
  /* All public properties are persisted */
  public {dollar_sign}id;
  /**
   * @var Programmer
   */
  public {dollar_sign}programmer;
  /**
   * @var Project
   */
  public {dollar_sign}project;
  public {dollar_sign}didProgrammerWin;
  public {dollar_sign}foughtAt;
  public {dollar_sign}notes;
}