Let's make the endpoint to create new battles. We're going to start like always by creating a new feature - battle.feature. The API clients are going to want to create battles to see if their programmers can take on and defeat these projects. After the business value, the next line is the person that's benefiting from the new feature and finally we have a little description.
Create Battle Scenario

Let's add the first Scenario: Creating a new Battle. If we go back to programmer.feature, we can copy a lot of this. First, in order to create a battle, we're probably going to need to be authenticated. So, I'll copy this background.

I'm going to go back and copy the entire scenario for creating a programmer. After all, this is an API, so creating a resource should always look pretty much the same.

Let's work from the end backwards and think about how we want the response to look. We know there's going to be a Location header, because there's always a Location header after creating a resource. But we don't know what URL that's going to be yet, because we don't have an endpoint yet for viewing a single battle. So we'll just say that the Location header should exist. And if you look at the Battle class, you'll see there's a didProgrammerWin property. Let's just make sure that exists as well - we don't know if it's going to be true or false, because there's some randomness. Let's update the URL to /api/battles and the status code of 201 looks perfect.

Creating a Programmer and Project First

In order to create a Battle - we'll need to send a programmer and a project. And probably the way we'll want the client to do that is by sending the programmer and the project's ids. So let's send programmerId and projectId - but we don't know yet what these should be set to.

Next, in order for us to start a battle, there needs to already be a programmer and a project sitting in the database. So before this line, we'll need to say Given there is a programmer called, and we'll create a new programmer called Fred. Again, these are all built-in Behat definitions that I created before we started working and they all live in either ApiFeatureContext or ProjectContext. If you want to know what I'm doing behind the scenes, just open up those classes. There's another one for And there is a project called, and we'll say 'my_project'