A little problem still exists: we don't know what the id's are of the programmer and project we just created. So I don't know what to put in the request body - we really want whatever ids those new things have. This is a really difficult problem with testing API's. So one of the things I've put into my testing system already, is the ability to do things like this.
It's a special syntax. And what this will do is go find a programmer whose nickname is 'Fred' and give us its id. It'll create a query for that dynamically. This syntax is totally special - it's not something built into Behat. If you want to know how it works, open ApiFeatureContext, scroll all the way to the bottom and find the processReplacements() function. It parses out that '%' syntax, looks for these wildcards, and lets us do some of that magic. This will be really handy, and we'll use it a few more times.

We'll do the same thing for projects. This looks great!

You know I like watching my tests fail first, so let's try it out. We'll just run this new battle.feature file:

php vendor/bin/behat features/api/battle.feature
Instead of 201, we get the 404 because the endpoint doesn't exist. That's awesome. Now let's make this work!