What's cool is that because we've done so much work up to this point, coding up our API is going to get really easy. I'll copy TokenController - I like having one resource per Controller. Update the class name. And this already has some code we want. So let's change the URL to /api/battles.
In newAction, we're going to reuse a lot of this. To create a Battle, you do need to be logged in, so we'll keep the enforceUserSecurity(). This decodeRequestBodyIntoParameters() is what goes out and reads the JSON on the request and gives us back this array-like object called a ParameterBag. So that's all good too.

I am going to remove most of the rest of this, because it's specific to creating a token.

Finding the Battle and Project

What we need to do is read the programmer and project id's off of the request and then create and save a new Battle object based off of those. So first, let's go get the projectId. We're able to use this nice get function, because the decodeRequestBodyIntoParameters function gives us that ParametersBag object. Let's also get the programmerId.

Perfect!

And with these 2 things, I need to query for the Project and Programmer objects, because the way I'm going to create the Battle will need them, not just their ids. Plus, this will tell us if these ids are even real. I'll use one of my shortcuts to query for the Project.

All this is doing is going out and finding the Project with that id and returning the Project model object that has the data from that row. We'll do the same thing with Programmer.