The BattleManager
Normally, to create a battle, you'd expect me to instantiate it manually. We've done that before for Tokens and Programmers. But for Battles, I have a helper called BattleManager that will do this for us. So instead of creating the Battle by hand, we'll call this battle() function and pass it the Programmer and Project:
58 lines  src/KnpU/CodeBattle/Battle/BattleManager.php
... lines 1 - 4
use KnpU\CodeBattle\Model\Battle;
use KnpU\CodeBattle\Model\Programmer;
use KnpU\CodeBattle\Model\Project;
... lines 8 - 10
class BattleManager
{
... lines 13 - 29
  public function battle(Programmer {dollar_sign}programmer, Project {dollar_sign}project){
    {dollar_sign}battle = new Battle();
    {dollar_sign}battle->programmer = {dollar_sign}programmer;
    {dollar_sign}battle->project = {dollar_sign}project;
    {dollar_sign}battle->foughtAt = new \DateTime();
    if ({dollar_sign}programmer->powerLevel < {dollar_sign}project->difficultyLevel) {
      // not enough energy
      {dollar_sign}battle->didProgrammerWin = false;
    } else {
      if (rand(0, 2) != 2) {
        {dollar_sign}battle->didProgrammerWin = true;
      } else {
        {dollar_sign}battle->didProgrammerWin = false;
        {dollar_sign}battle->notes = 'Requirements kept changing, too many meetings, project failed :(';
      }
    {dollar_sign}programmer->powerLevel = {dollar_sign}programmer->powerLevel - {dollar_sign}project->difficultyLevel;
    }
    {dollar_sign}this->battleRepository->save({dollar_sign}battle);
    {dollar_sign}this->programmerRepository->save({dollar_sign}programmer);
    return {dollar_sign}battle;
  }

It takes care of all of the details of creating the Battle, figuring out who won, setting the foughtAt time, adding some notes and saving all of this. So we do need to create a Battle object, but this will do it for us.

Creating the Battle

Back in BattleController, I already have a shortcut method setup to give us the BattleManager object. Then we'll use the battle() function we just saw and pass it the Programmer and Project.

And that's it - the Battle is created and saved for us. Now all we need to do is pass the Battle to the createApiResponse() method. And that will take care of the rest.

The createApiResponse method uses the serializer object to turn the Battle object into JSON. We haven't done any configuration on this class for the serializer, which means that it's serializing all of the fields. And for now, I'm happy with that - we're getting free functionality.

This looks good to me - so let's try it!

php vendor/bin/behat features/api/battle.feature
Oh! It almost passes. It gets the 201 status code, but it's missing the Location header. In the response, we can see the created Battle, with notes on why our programmer lost.