Adding the Location Header
Back in newAction, we can just set createApiResponse to a variable and then call {dollar_sign}response->headers->set() and pass it Location add a temporary todo.

Remember, this is the location to view a single battle, and we don't have an endpoint for that yet. But this will get our tests to pass for now:

php vendor/bin/behat features/api/battle.feature
Perfect!