Battle Validation
So let's add some validation. Since the battle() function is doing all of the work of creating the Battle, we don't need to worry about it too much. We just need to make sure that projectId and programmerId are valid. I'll just do validation manually here by creating an  variable and then check to see if we didn't find a Project in the database for some reason. If that's the case, let's add an error with a nice message. And we'll do the same thing with the Programmer.

Finally at the bottom, if we actually have at least one thing in the  variable, we're going to call a nice method we made in a previous chapter called throwApiProblemValidationException and just pass it the array of errors. It's just that easy.

We don't have a scenario setup for this, so let's tweak ours temporarily to try it - foobar is definitely not a valid id.

php vendor/bin/behat features/api/battle.feature:11
Now, we can see that the response code is 400 and we have this beautifully structured error response. So that's why we went to all of that work of setting up our error handling correctly, because the rest of our API is so easy and so consistent.