Next, let's keep going with viewing a single battle. Scenario: GETting a single battle. And thinking about this, we're going to need to make sure that there's a battle in the database first. I'm going to use similar language as before to create a Fred programmer and a project called project_facebook. I also have another step that allows me to say And there has been a battle between 'Fred' and 'project_facebook'. By the way, the nice auto-completion I'm getting is from the new PHPStorm 8 version, which has integration with Behat. I highly recommend it.
Great, so this makes sure there's something in the database. Next, we'll make the GET request to /api/battles/something. Here's the problem: the only way we can really identify our Battles are by their id. They're not like Programmer, where each has a unique nickname that we can use.

The Special %battles.last.id% Syntax

Here, we know there's a Battle in the database, but just like before when we were building the request body, we have no idea what that id was going to be. Fortunately, we can use that same magic % syntax. This time we can say %battles.last.id%

Before, we used this syntax to query for a Programmer by its nickname. But it also has a special 'last' keyword, which queries for the last record in the table. Again, this is me adding special things to my Behat project. Which is really handy for testing the API.

Next, go to programmer.feature and find its 'GET one programmer'. We'll copy the endpoint and 'Then' lines and do something similar. The status code looks good. The Battle has a didProgrammerWin field and we'll also make sure that the notes field is returned in the response.

You guys know the drill. We're going to try this first to make sure it fails. This is on line 26, so we'll add :26 to only run this scenario:

php vendor/bin/behat features/api/battle.feature:26
And there we go - we get the 404 instead of the 200 and that's perfect