Don't Forget to Fix the Location Header
Now that we have a proper showAction, we can go back and fix the todo in the header. First, we'll need to give the route an internal name - api_battle_show.
In newAction, we'll use generateUrl to make the URL for us.

Again, these shortcuts are things I added to my project, but this is just using the standard method in Silex to generate the URL based on the name of the route. And you can see what all of the shortcut methods really do by opening up the BaseController class.

First, let's make sure we didn't break anything by re-running the entire feature.

php vendor/bin/behat features/api/battle.feature
Green green green!