Let's do something really interesting. First I want to see what the Battle response looks like - so I'll say 'And print last response'
Now, run Behat.

php vendor/bin/behat features/api/battle.feature
At the bottom, you can see that response has programmer and project information right inside of it. That's because the Battle class has two properties that hold these objects. The serializer sees these objects, and serializes them recursively.

Controlling (Removing) Embedded Resources

In a couple of chapters, we're going to talk about embedding resources where we do this on purpose. But for now, I want to avoid it: if I'm getting a Battle, I only want to retrieve that Battle. So like before, we need to take control of how this is serialized.