I'll copy the Serializer use statement from Programmer into Battle. Next, let's also copy the ExclusionPolicy annotation into Battle, which tells the serializer to only serialize properties that we explicitly expose with the @Serializer\Expose annotation. Which properties you want to expose is totally up to you. I'll do it with the {dollar_sign}id, of course {dollar_sign}didProgrammerWin, {dollar_sign}foughtAt and we'll also expose the {dollar_sign}notes property.
You guys know the drill - let's run just line 26 to make sure things still pass:

php vendor/bin/behat features/api/battle.feature:26
We're still printing out the last response, but nothing is broken, so that's good. You can see that it's in fact not printing out the programmer or project anymore.

Adding a Link to the JSON Response

But now that we did that, if we think about somebody who is retrieving a single battle, they might want to know who the programmer was. I can hear our user now:

'ok, I see this battle, but what programmer fought in it and how can
get more information about them?'