So what I'll do is add a line to the scenario and look for a new, invented field: And the programmerUri field should equal '/api/programmers/Fred'
So we're saying that the response will have this extra field that's not really on the Battle class. What's cool about this is that as an API client, I'll see this and say:

'Oh, Fred was the programmer, and I can just go to that URL to get his details'.
First, let's run this and watch it fail:

php vendor/bin/behat features/api/battle.feature:26