Using a VirtualProperty
So how can we add this? The problem is that there really is no programmerUri property on Battle. So one of the cool features from JMS serializer is the ability to have virtual properties.

Create a new function called getProgrammerUri - the name of the method is important - and for right now, I'm just going to hardcode in the URL instead of generating it from the name like we have been doing. I'll fix that later.

But just because you have this method does not means it's going to be served out to your API. You can use an annotation called @Serializer\VirtualProperty.

And just like that, it's going to call getProgrammerUri, strip the get off of there, and look like a programmerUri field. And when I run my test, it does exactly that.

Congratulations! We just added our first link. And we're going to add a bunch more!