We just added a link to our API endpoint for getting a single battle. I want to see the response again, so let's add 'And print last response' then run the test - it starts on line 26.
php vendor/bin/behat features/api/battle.feature:26
We just invented this idea to put a field on a link called programmerUri. Part of the issue is that we have this link mixed up with other real data fields on this property. It's not totally obvious if we can follow this link, or if maybe this is just a field that happens to be a URL, and that we could actually change that URL by sending a PUT request if we wanted to.

HAL JSON: Standardizing How Links Look

A lot of smart people have thought about this and have invented different standardized formats for how your data and links should be organized inside of JSON or XML. One popular one right now is called HAL. I'll click into a document that has a really nice example:

{
  '_links': {
    'self': {
        'href': 'http://example.org/api/user/matthew'
    }
  }
  'id': 'matthew',
  'name': 'Matthew Weier O'Phinney'
}
You can see that down here is the data, and above that, you have a _links property that holds the links. You'll also notice that there's this link called self, and that's something we're going to see over and over again. self is kind of a standard thing where each resource has a link to itself, and you keep that on a key called self. What's cool about this is that it's used on a lot of APIs. So if you ever see a link with the key self, you already know what they're talking about. We're going to add this to our stuff too.

Right now, we're using the serializer to create our JSON. And it works just by looking at the class of whatever object we're serializing and grabs the properties off of it. And of course we have some control over which properties to use.

But if you look at the _links.self thing, you might be wondering how we're going to do this. Are we going to need a bunch of these VirtualProperty things for _links?