How HATEOAS Works
Just like the serializer, this library works via annotations:

use Hateoas\Configuration\Annotation as Hateoas;

/**
 * @Hateoas\Relation('self', href = 'expr('/api/users/' ~ object.getId())')
 */
class User {
  private {dollar_sign}id;
  private {dollar_sign}firstName;
  private {dollar_sign}lastName;

  public function getId() {}
}
It basically says the User class has a relation called self, and its href should be set to /api/users, and then the id of the user. And we'll talk about this syntax in a second. The end result will be something that looks like this:

{
  '_links': {
    'self': {
      'href': '/api/users/12'
    }
  }
    'id': '12',
    'firstName': 'Leanna',
    'lastName': 'Pelham'
}
It'll create the _links key with self and any other links you add below it.

HATEOAS Setup

Getting this setup is pretty easy. Find the HateoasBuilder code and copy it. Quickly, I'll make sure Composer is done downloading the library.

There's always a place inside a Silex application - and most frameworks - to define services, which are just re-usable objects. You might remember from earlier, that in my project, this is done inside this Application class. A few chapters ago, we used this to configure the serializer object. The HATEOAS library hooks right into this, so we just need to modify a few things. Instead of returning the serializer, we'll set it to a variable and take off the call to build(). Now we'll paste the new code in, return it, and pass the  into the create() method, which is what ties the two libraries together.

Now of course, my editor is angry because I'm missing my use statement, so I'll use a shortcut to add that, which puts use statement at the top of this file.

So that's it. We're already using the serializer object everywhere, and now the HATEOAS library will be working with that to add these links for us.