Update the Scenario for Links
Before we add our first link, let's add a scenario to look for it. Go into programmer.feature and find the scenario for getting one programmer. As I mentioned before. it's always a really good idea to have a self link. And if we look at the structure of HAL, this means we're going to have a _links.self.href property, and it'll be set to the URI of our programmer.

So let's add that here: And the '_links.self.href' property should equal - and we know what the URL of this is going to be - /api/programmers/UnitTester.

And as always, let's run this to see it fail. This scenario is on line 66 of programmer.feature:

php vendor/bin/behat features/api/programmer.feature:66
And it fails because the property doesn't even exist yet.