This is great, except that I don't like this hardcoded URL in the Relation.
54 lines  src/KnpU/CodeBattle/Model/Programmer.php
... lines 1 - 7
use Hateoas\Configuration\Annotation as Hateoas;
... line 9
/**
 * @Serializer\ExclusionPolicy('all')
 * @Hateoas\Relation('self', href = 'expr('/api/programmers/' ~ object.nickname)')
 */
class Programmer
{
... lines 16 - 52
}

What I'd rather do is generate that URL from the internal name api_programmers_show like we normally do.

Fortunately, the HATEOAS library allows us to do that(http://hateoas-php.org/#url-generators), and this can be hooked up to work with any framework, since they all generate URLs. As part of that HateoasBuilder::create() step, you can set a URL generator that does whatever you need.

Hooking up the SymfonyUrlGenerator

If you're using Silex or Symfony, life is a little bit easier because there's a built-in class called SymfonyUrlGenerator:
// from the HATEOAS docs
use Hateoas\UrlGenerator\SymfonyUrlGenerator;

 = HateoasBuilder::create()
    ->setUrlGenerator(null, new SymfonyUrlGenerator(['url_generator']))
    ->build()
;

I'll copy this line. Go back into Application, move build() onto the next line and paste this

And don't forget that we need a use statement for that SymfonyUrlGenerator, so I'll click 'import' to have PhpStorm add this class to the top of the file for me. This class comes from the HATEOAS library, and we're just passing it the url_generator object, which in Silex, is the object responsible for generating URLs. In Symfony, it's called router.