Consistency = New Behat Step
Because we're always putting links under a _links key, I have a new piece of language that we can use in Behat to check for links.

Why would I do this? It's just proving how consistent we are. This new sentence will look for the _links property, so there's no reason to repeat it in all of our scenarios. So let's try the test again - perfect.

php vendor/bin/behat features/api/battle.feature:26
We can repeat this same thing over in programmer.feature when we're checking the self link. I'll comment out the old line for reference.

If we run our entire test suite, things keep passing:

php vendor/bin/behat
I love to see all of that green!