Embedding a Resource
So let's try to embed our programmer into the battle. First, let's add a line to the scenario that looks for this. Let's look for _embedded, and we know it's going to be called programmer, and the data will live below this and we know one of the fields on a programmer is nickname. And we know this should be equal to Fred.

Let's make sure that fails first - and it does:

php vendor/bin/behat features/api/battle.feature:26